// data/nepse/types.ts
export type Logger = {
  info: (msg: string, ...args: unknown[]) => void;
  warn: (msg: string, ...args: unknown[]) => void;
  error: (msg: string, ...args: unknown[]) => void;
};

export type ClientState = {
  nepseExports: NepseExports | null;
  token: {
    value: string | null;
    expiry: number;
  };
  caches: {
    securityBriefs: Cache<SecurityBrief[]>;
  };
  isInitialized: boolean;
  logger?: Logger;
};
export type MarketStatus = {
  isOpen: string;
  asOf: string;
  id: number;
};

export type Prove = {
  accessToken: string;
  refreshToken: string;
  salt: string;
  salt1: number;
  salt2: number;
  salt3: number;
  salt4: number;
  salt5: number;
  serverTime: number;
  tokenType: string | null;
  isDisplayActive: boolean;
  popupDocFor: string;
};

export type ProveExtra = {
  serverTime: number;
  clientTime: number;
  validAccessToken: string;
};

export type SecurityBrief = {
  activeStatus: string;
  id: number;
  name: string;
  securityName: string;
  symbol: string;
};

export type SecurityDetail = {
  id: number;
  symbol: string;
  name: string;
  activeStatus: string;
  listingDate: Date;
  closePrice: number;
  businessDate: Date;
  fiftyTwoWeekHigh: number;
  fiftyTwoWeekLow: number;
  lastTradePrice: number;
};
// ...
export type IndexDetail = {
  id: number;
  auditId: any;
  exchangeIndexId: any;
  generatedTime: string;
  index: string;
  close: number;
  high: number;
  low: number;
  previousClose: number;
  change: number;
  perChange: number;
  fiftyTwoWeekHigh: number;
  fiftyTwoWeekLow: number;
  currentValue: number;
};

export type NepseDecodeFunction = (
  var0: number,
  var1: number,
  var2: number,
  var3: number,
  var4: number,
) => number;

export type NepseExports = {
  cdx: NepseDecodeFunction;
  rdx: NepseDecodeFunction;
  bdx: NepseDecodeFunction;
  ndx: NepseDecodeFunction;
  mdx: NepseDecodeFunction;
};

export type NepseError = Error & {
  code: string;
  originalError?: unknown;
};

export type CacheItem<T> = {
  data: T;
  expiry: number;
};

export type Cache<T> = {
  [key: string]: CacheItem<T>;
};

// ...existing code...

export type InitializeOptions = {
  useWasm?: boolean;
  logger?: Logger;
};